jack3/jack3/requirements.txt
# Jack v3 — BankNifty AI Intraday Trader
# Install with: pip install -r requirements.txt

# Core data science
pandas>=2.0
numpy>=1.24

# Configuration
pyyaml>=6.0

# HTTP requests (for FII/DII scraping, openai proxy)
requests>=2.28
beautifulsoup4>=4.11

# Market data (free, no API key needed)
yfinance>=0.2

# AI providers (install the one(s) you need)
anthropic>=0.40    # For Anthropic API backend

# Broker SDK — ONLY install when ready to go live
# dhanhq>=1.4     # Uncomment when trading live: pip install dhanhq

# Optional: OpenAI SDK (not needed if using raw requests for the proxy)
# openai>=1.0
